chore: created generic API hook and GameCollection implementation
