+++ src/hooks/useGamesCollection.ts
@@ -1,0 +1,8 @@
+import useAPICollection from "./useAPICollection"
+
+interface Game {
+    id: number,
+    name: string
+}
+
+export default () => useAPICollection<Game>('/games');
